+++ bin/lib/pan-python/lib/pan/config.py
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2012-2014 Kevin Steves <kevin.steves@pobox.com>
+# Copyright (c) 2012-2017 Kevin Steves <kevin.steves@pobox.com>
 #
 # Permission to use, copy, modify, and distribute this software for any
 # purpose with or without fee is hereby granted, provided that the above
@@ -254,8 +254,10 @@ class PanConfig:
         for e in elem:
             self.__serialize_flat(e, path + '/' + e.tag, obj)
 
-    def __quote_space(self, s):
+    def __quote_arg(self, s):
         # XXX string with " etc.
+        if '"' in s:
+            return "'%s'" % s
         if ' ' in s:
             return '"%s"' % s
         return s
@@ -291,7 +293,7 @@ class PanConfig:
 
         for k, v in attrs:
             if k == 'name':
-                path += ' ' + self.__quote_space(v)
+                path += ' ' + self.__quote_arg(v)
 
         if member_list:
             nodes = elem.findall('./member')
@@ -299,14 +301,14 @@ class PanConfig:
             if len(nodes) > 1:
                 members = []
                 for e in nodes:
-                    members.append(self.__quote_space(e.text))
+                    members.append(self.__quote_arg(e.text))
                 path += ' [ ' + ' '.join(members) + ' ]'
                 obj.append(path)
                 return
 
         if not len(elem):
             if text_strip:
-                path += ' ' + self.__quote_space(text)
+                path += ' ' + self.__quote_arg(text)
             obj.append(path)
 
         for e in elem:
@@ -430,6 +432,60 @@ class PanConfig:
 ./mgt-config
 '''
 
+        xpaths_panos_7_0 = '''
+./devices/entry[@name='localhost.localdomain']/deviceconfig
+./devices/entry[@name='localhost.localdomain']/network
+./shared
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/zone
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/vm-info-source
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/user-id-collector
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/user-id-agent-sequence
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/user-id-agent
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/url-content-types
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/url-admin-override
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/ts-agent
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/threats
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/tag
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/setting
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/service-group
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/service
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/schedule
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/rulebase
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/route
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/reports
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/report-group
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/region
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/profiles
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/profile-group
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/pdf-summary-report
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/import
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/group-mapping
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/global-protect
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/external-list
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/email-scheduler
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/dns-proxy
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/display-name
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/captive-portal
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/application-group
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/application-filter
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/application
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/address-group
+./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/address
+./mgt-config
+'''
+
+        s = '''./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/application-tag'''
+        x = xpaths_panos_7_0.split('\n')
+        x.insert(34, s)
+        xpaths_panos_7_1 = '\n'.join(x)
+
+        s = '''./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/user-id-ssl-auth'''
+        x = xpaths_panos_7_1.split('\n')
+        x.insert(6, s)
+        s = '''./devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/authentication-object'''
+        x.insert(35, s)
+        xpaths_panos_8_0 = '\n'.join(x)
+
         xpaths_panos_multi_vsys_4_1 = '''
 ./devices/entry[@name='localhost.localdomain']/deviceconfig
 ./devices/entry[@name='localhost.localdomain']/network
@@ -437,6 +493,7 @@ class PanConfig:
 ./devices/entry[@name='localhost.localdomain']/vsys
 ./mgt-config
 '''
+
         xpaths_panorama_4_1 = '''
 ./devices/entry[@name='localhost.localdomain']/deviceconfig
 ./devices/entry[@name='localhost.localdomain']/device-group
@@ -483,6 +540,40 @@ class PanConfig:
 ./mgt-config
 '''
 
+        # reorder
+        # add readonly
+        xpaths_panorama_7_0 = '''
+./devices/entry[@name='localhost.localdomain']/device-group
+./devices/entry[@name='localhost.localdomain']/deviceconfig
+./devices/entry[@name='localhost.localdomain']/log-collector
+./devices/entry[@name='localhost.localdomain']/log-collector-group
+./devices/entry[@name='localhost.localdomain']/template
+./devices/entry[@name='localhost.localdomain']/template-stack
+./devices/entry[@name='localhost.localdomain']/vmware-service-manager
+./mgt-config
+./panorama
+./readonly
+./shared
+'''
+
+        # remove vmware-service-manager
+        # add plugins, wildfire-appliance, wildfire-appliance-cluster
+        xpaths_panorama_8_0 = '''
+./devices/entry[@name='localhost.localdomain']/device-group
+./devices/entry[@name='localhost.localdomain']/deviceconfig
+./devices/entry[@name='localhost.localdomain']/log-collector
+./devices/entry[@name='localhost.localdomain']/log-collector-group
+./devices/entry[@name='localhost.localdomain']/plugins
+./devices/entry[@name='localhost.localdomain']/template
+./devices/entry[@name='localhost.localdomain']/template-stack
+./devices/entry[@name='localhost.localdomain']/wildfire-appliance
+./devices/entry[@name='localhost.localdomain']/wildfire-appliance-cluster
+./mgt-config
+./panorama
+./readonly
+./shared
+'''
+
         xpaths_panos = xpaths_panos_4_1
         xpaths_panos_multi_vsys = xpaths_panos_multi_vsys_4_1
         xpaths_panorama = xpaths_panorama_4_1
@@ -497,6 +588,15 @@ class PanConfig:
             elif self.config_version() in ['6.1.0']:
                 xpaths_panos = xpaths_panos_6_0
                 xpaths_panorama = xpaths_panorama_6_1
+            elif self.config_version() in ['7.0.0']:
+                xpaths_panos = xpaths_panos_7_0
+                xpaths_panorama = xpaths_panorama_7_0
+            elif self.config_version() in ['7.1.0']:
+                xpaths_panos = xpaths_panos_7_1
+                xpaths_panorama = xpaths_panorama_7_0
+            elif self.config_version() in ['8.0.0']:
+                xpaths_panos = xpaths_panos_8_0
+                xpaths_panorama = xpaths_panorama_8_0
 
         if self.config_multi_vsys():
             xpaths = xpaths_panos_multi_vsys
@@ -504,7 +604,10 @@ class PanConfig:
             xpaths = xpaths_panorama
         else:
             xpaths = xpaths_panos
+
+        self._log(DEBUG2, '%s', xpaths)
         xpaths = xpaths.split('\n')
         xpaths = [s for s in xpaths if s]
+        self._log(DEBUG1, 'xpaths: %d', len(xpaths))
 
         return xpaths