+++ bin/lib/pan-python/lib/pan/xapi.py
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2013-2015 Kevin Steves <kevin.steves@pobox.com>
+# Copyright (c) 2013-2016 Kevin Steves <kevin.steves@pobox.com>
 #
 # Permission to use, copy, modify, and distribute this software for any
 # purpose with or without fee is hereby granted, provided that the above
@@ -74,7 +74,8 @@ class PanXapi:
                  use_http=False,
                  use_get=False,
                  timeout=None,
-                 ssl_context=None):
+                 ssl_context=None,
+                 **kwargs):
         self._log = logging.getLogger(__name__).log
         self.tag = tag
         self.api_username = None
@@ -86,9 +87,12 @@ class PanXapi:
         self.use_get = use_get
         self.timeout = timeout
         self.ssl_context = ssl_context
+        self._legacy_api = kwargs.get('_legacy_api', False)
 
         self._log(DEBUG3, 'Python version: %s', sys.version)
         self._log(DEBUG3, 'xml.etree.ElementTree version: %s', etree.VERSION)
+        if hasattr(ssl, 'OPENSSL_VERSION'):  # added 2.7
+            self._log(DEBUG3, 'ssl: %s', ssl.OPENSSL_VERSION)
         self._log(DEBUG3, 'pan-python version: %s', __version__)
 
         if self.port is not None:
@@ -112,6 +116,12 @@ class PanXapi:
                 ssl.SSLContext(ssl.PROTOCOL_SSLv23)
             except AttributeError:
                 raise PanXapiError('SSL module has no SSLContext()')
+
+        # handle Python versions with no ssl.CertificateError
+        if hasattr(ssl, 'CertificateError'):
+            self._certificateerror = ssl.CertificateError
+        else:
+            self._certificateerror = NotImplementedError  # XXX Can't happen
 
         init_panrc = {}  # .panrc args from constructor
         if api_username is not None:
@@ -179,7 +189,8 @@ class PanXapi:
         self.uri = '%s://%s' % (scheme, self.hostname)
         if self.port is not None:
             self.uri += ':%s' % self.port
-        self.uri += '/api/'
+        # _legacy_api is used for PAN-OS < 4.1.0
+        self.uri += '/api/' if not self._legacy_api else '/esp/restapi.esp'
 
         if _legacy_urllib:
             self._log(DEBUG2, 'using legacy urllib')
@@ -534,11 +545,15 @@ class PanXapi:
             response = urlopen(**kwargs)
 
         # XXX handle httplib.BadStatusLine when http to port 443
+        except self._certificateerror as e:
+            self.status_detail = 'ssl.CertificateError: %s' % e
+            return False
         except URLError as error:
             msg = 'URLError:'
             if hasattr(error, 'code'):
                 msg += ' code: %s' % error.code
             if hasattr(error, 'reason'):
+                self._rome_warning(str(error.reason))
                 msg += ' reason: %s' % error.reason
             if not (hasattr(error, 'code') or hasattr(error, 'reason')):
                 msg += ' unknown error (Kevin heart Python)'
@@ -549,6 +564,19 @@ class PanXapi:
         self._log(DEBUG2, '%s', response.info())
 
         return response
+
+    def _rome_warning(self, reason):
+        # "Connection reset by peer" or
+        # "EOF occurred in violation of protocol"
+        # may indicate no TLS 1.[12]
+        if (('Errno 54' in reason or
+                'EOF occurred in violation of protocol' in reason) and
+                hasattr(ssl, 'OPENSSL_VERSION_NUMBER') and
+                ssl.OPENSSL_VERSION_NUMBER < 0x1000100):
+            x = 'WARNING: Your SSL (%s) may not support TLS 1.1.' % \
+                ssl.OPENSSL_VERSION
+            x += ' PAN-OS 8.0 does not allow TLS 1.0 connections by default.'
+            self._log(DEBUG1, x)
 
     def __set_api_key(self):
         if self.api_key is None:
@@ -799,6 +827,36 @@ class PanXapi:
         if not self.__set_response(response):
             raise PanXapiError(self.status_detail)
 
+    def __legacy_commit_poll(self, query, interval=None, timeout=None):
+        """On PAN-OS < 4.1.0, keep polling until the commit is finished.
+
+        type=op API request is not available in PAN-OS < 4.1.0, so
+        instead of checking on the job id, keep trying to commit until
+        there is nothing to commit.
+        """
+
+        start_time = time.time()
+        while True:
+            # sleep at the top of the loop so we don't poll
+            # immediately after commit
+            self._log(DEBUG2, 'sleep %.2f seconds', interval)
+            time.sleep(interval)
+
+            response = self.__api_request(query)
+            if not response:
+                raise PanXapiError(self.status_detail)
+            self.__set_response(response)
+            if self.status_detail == "There are no changes to commit.":
+                self._log(DEBUG2, 'commit finished')
+                return
+            else:
+                self._log(DEBUG2, 'commit pending...')
+
+            if (timeout is not None and timeout != 0 and
+                    time.time() > start_time + timeout):
+                raise PanXapiError('timeout waiting for legacy commit ' +
+                                   'completion')
+
     def commit(self, cmd=None, action=None, sync=False,
                interval=None, timeout=None, extra_qs=None):
         self.__set_api_key()
@@ -833,7 +891,9 @@ class PanXapi:
             query['action'] = action
         if extra_qs is not None:
             query = self.__merge_extra_qs(query, extra_qs)
-
+        if self._legacy_api:
+            query['type'] = 'config'
+            query['action'] = 'commit'
         response = self.__api_request(query)
         if not response:
             raise PanXapiError(self.status_detail)
@@ -843,6 +903,11 @@ class PanXapi:
 
         if sync is not True:
             return
+
+        if self._legacy_api:
+            return self.__legacy_commit_poll(query,
+                                             interval=interval,
+                                             timeout=timeout)
 
         job = self.element_root.find('./result/job')
         if job is None: